backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.2.0
pytesseract==0.3.10
SpeechRecognition==3.10.1
pydub==0.25.1
